Add notifications feature and related UI components (#62)
* Add notifications feature

Several new files and modifications have been introduced to set up the notifications feature. This includes the creation of notification services, schemas, and requests, as well as modifying the main layout to include notifications. The notifications view and UI were established and a connection hub for notifications was formed.

* Update notification hooks and display

The `useNotifications` hook was updated for a better handling of notifications data. Notification display was also refined in `NotificationPage` and `Notifications` components to cater to the changes. Additionally, some minor UI adjustments and improvements in labelling were made within various files.

* Рефакторинг